% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_layers.R
\name{filter_county_poly}
\alias{filter_county_poly}
\title{Retrieve county polygon layer}
\usage{
filter_county_poly(...)
}
\arguments{
\item{...}{Any Wisconsin counties provided as character strings, separated
by commas}
}
\value{
An sf data.frame with the appropriate counties
}
\description{
Return specific county polygon layer from \code{wi_counties} sf object
}
\examples{
\dontrun{
plot(filter_county_poly("door"))
plot_layer(filter_county_poly("portage"))
}
}
