% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{list_layer_url}
\alias{list_layer_url}
\title{Helper function to re-create \code{\link{list_layers}} and
\code{\link{list_urls}}}
\usage{
list_layer_url(type = "layer", sections = NULL, services = NULL, pull = TRUE)
}
\arguments{
\item{type}{Character. The column of data to retrieve from service_urls}

\item{sections}{See \code{\link{list_funs}}}

\item{services}{See \code{\link{list_funs}}}

\item{pull}{See \code{\link{list_funs}}}
}
\value{
A vector of available layers or URLs; depending on \code{type}
}
\description{
Helper function to re-create \code{\link{list_layers}} and
\code{\link{list_urls}}
}
