% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_layer_functions.R
\name{match_watershed_name}
\alias{match_watershed_name}
\title{Match a watershed's name based on one or more regex}
\usage{
match_watershed_name(..., pull = TRUE)
}
\arguments{
\item{...}{One or more regex passed as character string}

\item{pull}{Logical. Pull the unique values or}
}
\value{
A character string with full watershed names if pull = TRUE, or a
data.frame with the number of rows equal to the number of matches otherwise
}
\description{
This function will match the names of a HUC_8 or a HUC_12 watershed found in
the \code{watershed_lookup} data set.
}
\examples{
match_watershed_name("rainbow")
}
