% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{standardize_county_names}
\alias{standardize_county_names}
\title{Standardize county names}
\usage{
standardize_county_names(...)
}
\arguments{
\item{...}{One or more county names in quotations, or a character vector of
county names}
}
\value{
A character vector the same length as \code{name}, but tidied up
for easier and standard viewing
}
\description{
This function alters string text of county names to a standardized format of
lower-cased, no punctuation (i.e. st instead of st.), and underscore instead
of spaces
}
