% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wi_example_data}
\alias{wi_example_data}
\alias{service_urls}
\title{Various example data and lookup tables}
\format{
A data.frame
}
\usage{
service_urls
}
\description{
These datasets are used for functions and examples throughout the package
}
\keyword{datasets}
