% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R, R/sf_polygon_data.R
\docType{data}
\name{watershed_lookup}
\alias{watershed_lookup}
\alias{wi_example_polys}
\alias{wi_counties}
\alias{wi_poly}
\title{Various example sf polygons}
\format{
An object of class \code{data.frame} with 2232 rows and 3 columns.

An object of class sf and data.frame:

An object of class \code{sf} (inherits from \code{data.frame}) with 1 rows and 2 columns.
}
\source{
\code{\link[ggplot2]{map_data}}
}
\usage{
watershed_lookup

wi_counties

wi_poly
}
\description{
These are sf polygons that are used for functions and examples
throughout the package
}
\keyword{datasets}
